requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.3.6
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
